Gene Ontology Query
Source: R/getGO.R

This function takes a gene id as character vector from a given query organism and returns the corresponding GO terms and additional GO information.

getGO(organism, genes, filters, ...)

Arguments

organism
a character string specifying the scientific name of a query organism.

genes
a character vector storing the gene ids of a organisms of interest to be queried against Ensembl Biomart.

filters
a character vector specifying the filter (query key) for the Ensembl Biomart query, e.g. filter = "ensembl_gene_id".

...
additional parameters that can be passed to the biomart function.

Details

This function takes the scientific name of a query organism, a set of genes for which GO terms and additional information shall be retrieved, and a filter argument that specifies the attribute for the query genes.

See also

Examples

# NOT RUN {
GO_tbl <- getGO(organism = "Arabidopsis thaliana",
                genes    = c("AT1G06090", "AT1G06100"),
                filters  = "ensembl_gene_id")

# look at the result
head(GO_tbl)
# }