Retrieve metagenomes from NCBI Genbank

Retrieve available metagenomes from NCBI Genbank. NCBI Genbank allows users to download entire metagenomes of several metagenome projects. This function downloads available metagenomes that can then be downloaded via getMetaGenomes.

getMetaGenomes(
  name,
  path = file.path("_ncbi_downloads", "metagenome"),
  metagenomes.members = dplyr::filter(getMetaGenomeSummary(), organism_name == name)
)

Arguments

name

metagenome name retrieved by listMetaGenomes.

path

a character string specifying the location (a folder) in which the corresponding metagenome shall be stored. Default is path = file.path("_ncbi_downloads","metagenome").

metagenomes.members

a tibble of metagenome assemblies, default: dplyr::filter(getMetaGenomeSummary(), organism_name == name)

See also

Author

Hajk-Georg Drost

Examples

if (FALSE) { # \dontrun{
# Frist, retrieve a list of available metagenomes
listMetaGenomes()

# Now, retrieve the 'human gut metagenome'
getMetaGenomes(name = "human gut metagenome")
} # }