Import GFF File

This function reads an organism specific CDS stored in a defined file format.

read_gff(file)

Arguments

file

a character string specifying the path to the file storing the CDS.

Value

Either a Biostrings or data.table object.

Details

This function takes a string specifying the path to the GFF file of interest (e.g. the path returned by getGFF).

See also

Author

Hajk-Georg Drost